backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
SQLAlchemy>=1.4
pydantic>=1.10
email-validator>=1.1.3
python-multipart>=0.0.5
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
boto3>=1.26.0
python-dotenv>=1.0.0
